move methods related to setting context_metadata types for BindingIdentifiers and Labels
